requirements.txt
openai>=1.0.0
anthropic>=0.18.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
